requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-json-logger==2.0.7
sentry-sdk[flask]==1.40.4 